upload script via USB to Victron GX

+++ venus-data/module_m/install.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 echo "Installing Home Manager dbus bridge..."
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+# SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+SCRIPT_DIR="/data/module_m"
+
+# wait until dbus settings are active (when installing via blind install)
+while [ $(dbus -y | grep -c "com.victronenergy.settings") == 0 ]; do
+    logMessage "waiting for dBus settings"
+    sleep 1
+done
+
+sleep 2
 
 echo "Generate start and stop scripts.."
 # Kill script
@@ -24,6 +33,10 @@ if [ ! -f "/data/rc.local" ]; then
   echo "#!/bin/bash" > /data/rc.local
   chmod +x /data/rc.local
 fi
-echo "ln -s $SCRIPT_DIR/service /service/venus-homemanager" >> /data/rc.local
+
+# check if rc.local already contains the service
+if [ $(grep -c "ln -s $SCRIPT_DIR/service /service/venus-homemanager" /data/rc.local) -eq 0 ]; then
+  echo "ln -s $SCRIPT_DIR/service /service/venus-homemanager" >> /data/rc.local
+fi
 
 echo "Installation finished!"
